Crate tlist

An indexable list structure based on a Red-Black tree that uses sub-tree counts as a 'key' to achieve O(lg N) insertion, deletion, and access. Originally designed for rapid merge operations of a list with a sequence of indexed deltas.

The key difference in the TList from a standard Red-Black tree, and in particular from using a RBT, e.g. using a usize 'index' as a key, is that the index value is implicit in TList rather than explicit like a normal key in a BST. If the index were used as an explicit key, then all nodes to the right of the insertion location would need their key adjusted upward by one, essentially reducing the RBT to an array with O(lg N) search, and the same big-O behavior as an array for other operations. This is strictly worse than an array or vector.

Expressing the index implicity allows TList to act like an Array or Vector with O(lg N) random insert and delete at the cost of O(lg N) random-access instead of O(1). This is more efficient in situations where reads are primarily via iteration over the whole collection, or ranges, but mutation is primarily through insert/delete, e.g. delta-merge operations.

Implemented internally using a Vec and indexes in an Arena-like fashion.

Structs

IntoIter
An iterator over a TList that consumes the list and yields owned values, equivalent to an IntoIter over a Vec.

Iter
An iterator over a TList that yields borrowed values in index-order, equivalent to an Iter over a Vec.

TList
An indexable list structure based on a Red-Black tree that uses sub-tree counts as a 'key' to achieve O(lg N) insertion, deletion, and access. Originally designed for rapid merge operations of a list with a sequence of indexed deltas.