Struct tlist::TList

pub struct TList<T: Sized> { /* fields omitted */ }

An indexable list structure based on a Red-Black tree that uses sub-tree counts as a 'key' to achieve O(lg N) insertion, deletion, and access. Originally designed for rapid merge operations of a list with a sequence of indexed deltas.

The index is expressed implicity, which allows TList to act like an Array or Vector with O(lg N) random insert and delete at the cost of O(lg N) random-access instead of O(1). This is more efficient in situations where reads are primarily via iteration over the whole collection, or ranges, but mutation is primarily through insert/delete, e.g. delta-merge operations.

Implemented internally using a Vec and indexes in an Arena-like fashion.

Methods

impl<T> TList<T> where T: Sized

fn new() -> TList<T>

Generate a new TList with the default capacity.

fn with_capacity(capacity: usize) -> TList<T>

Initialize a new TList with a chosen capacity

fn len(&self) -> usize

Gets the number of elements in the TList

fn capacity(&self) -> usize

Gets the total possible number of elements the TList can store without reallocating.

fn from_data<U>(data: &[U]) -> TList<U> where U: Sized + Clone

Build a new TList from a vector of data elements, such that the inorder traversal of the TList maintains the same ordering as the original vector.

fn insert(&mut self, elem: T, index: usize)

Insert a new value into the TList at the given index. Unlike a regular Vec, insertion at one past the final element (the TList.len() value) is valid and equivalent to push(). Insertion is guarenteed to operate in O(log N).

Note: Insertion at an index past the len() value of the list currently appends, but this behavior is not guarenteed by the contract of this function and may change. Use insert_or_push() which specifies that behavior.

fn push(&mut self, elem: T)

Inserts an element at the end of the list. Unlike Vec, this is an O(log N) operation.

fn insert_or_push(&mut self, elem: T, index: usize)

Inserts an element at index if the index is valid, and appends the element per push if the index is invalid past the end.

fn get(&self, index: usize) -> Option<&T>

Returns a reference to the value at index if it exists. Runs in O(log N) time.

fn get_mut(&mut self, index: usize) -> Option<&mut T>

Returns a mutable reference to the value at index if it exists. Runs in O(log N) time.

fn remove(&mut self, index: usize) -> Option<T>

Deletes the value at the specified index, returning an owned value if it exists. All higher indexes are shifted down "implicitly", yielding the same behavior as removing a value from a Vec but with O(log N) running time.

fn pop(&mut self) -> Option<T>

Equivalent to remove() on the last element of the TList.

fn iter(&self) -> Iter<T>

Returns an iterator over the values in the TList in index order, yielding immutable borrows. Mirrors Vec's iter() and Iter.

fn into_iter(self) -> IntoIter<T>

Consumes self and returns an iterator over the values in the TList in index order, yielding owned values. Mirrors Vec's into_iter() and IntoIter.